% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{all_models_in_dir}
\alias{all_models_in_dir}
\title{Extract names of all models in Examples}
\usage{
all_models_in_dir(examples_dir)
}
\arguments{
\item{examples_dir}{A character vector (length 1), containing the path to
the Examples directory in the MultiBUGS directory}
}
\value{
A character vector of model names
}
\description{
The Examples in BUGS are named in a standard manner. For a model called
"EXAMPLE", there will be the following files:
}
\details{
1. EXAMPLE.txt - this is the manual page that describes the model, and links
   to other pages
2. EXAMPLEmodel.txt - this just contains the model
3. EXAMPLEdata.txt - this just contains the data
4. EXAMPLEinits.txt - this contains the first set of inits
5. EXAMPLEinits1.txt - this contains a second set of inits

This function uses 2 to extract a list of all models in the Examples
directory
}
