% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appveyor.R
\name{appveyor_reporter}
\alias{appveyor_reporter}
\title{Report result of test for Appveyor}
\usage{
appveyor_reporter(type, ...)
}
\arguments{
\item{type}{One of \code{"setup"}, \code{"pre"}, \code{"post"} and
\code{"wrapup"} specifying the stage that needs reporting.}
}
\value{
The function \code{\link{appveyor_reporter_setup}},
\code{\link{appveyor_reporter_pre}}, \code{\link{appveyor_reporter_post}}
or \code{\link{appveyor_reporter_wrapup}}
}
\description{
Report result of test for Appveyor
}
