% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appveyor.R
\name{appveyor_reporter_post}
\alias{appveyor_reporter_post}
\title{Report "post" results to Appveyor}
\usage{
appveyor_reporter_post(fit, problem_table_string, passed, model, n.workers,
  milliseconds, working.directory)
}
\arguments{
\item{fit}{The output of the test. This should be an object for which
\code{\link{print}} will appropriately summarise}

\item{passed}{A logical vector (length 1) summarising the result of the
test}

\item{model}{A character vector (length 2) containing the name of the model
being tested}

\item{n.workers}{The number of workers used}

\item{milliseconds}{The number of milliseconds that the test took}

\item{working.directory}{Path to the working directory used for the run}
}
\description{
Report "post" results to Appveyor
}
