% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appveyor.R
\name{appveyor_reporter_pre}
\alias{appveyor_reporter_pre}
\title{Report "pre" results to AppVeyor}
\usage{
appveyor_reporter_pre(model, n.iter, n.chains, n.workers,
  working.directory)
}
\arguments{
\item{model}{A character vector (length 2) containing the name of the model
being tested}

\item{n.workers}{The number of workers used}

\item{working.directory}{Path to the working directory used for the run}
}
\description{
Report "pre" results to AppVeyor
}
