% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appveyor.R
\name{appveyor_reporter_setup}
\alias{appveyor_reporter_setup}
\title{Report "setup" results to AppVeyor}
\usage{
appveyor_reporter_setup(dir, n.chains, n.workers, examples_dir)
}
\arguments{
\item{n.workers}{The number of workers used}
}
\description{
Report "setup" results to AppVeyor
}
