% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{bugs_example}
\alias{bugs_example}
\title{Run a single Example}
\usage{
bugs_example(model, n.iter = NULL, n.chains = 2, n.workers = 2,
  dir = "C:/MultiBUGS", working_dir = tempdir(),
  implementation = "MultiBUGS", examples.dir = file.path(dir,
  "Examples"))
}
\arguments{
\item{model}{A character vector (length 1) specifying the model}

\item{n.iter}{The number of iterations. If NULL, the default for each model
is used.}

\item{n.chains}{The number of chains}

\item{n.workers}{The number of cores to use}

\item{dir}{Full path to the MultiBUGS install directory}

\item{working_dir}{Full path to a temp dir where the model should be run}

\item{implementation}{Either \code{"MultiBUGS"} or \code{"OpenBUGS"}}

\item{examples.dir}{Path to the directory containing the examples}
}
\description{
Runs a single Example from the specified MultiBUGS install, with the
specified level of distribution
}
