% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{bugs_example_compare}
\alias{bugs_example_compare}
\title{Run a single Example}
\usage{
bugs_example_compare(multidir = NULL,
  opendir = "S:/winapps/openbugs/OpenBUGS323-with-MultiBUGS-examples",
  ...)
}
\arguments{
\item{multidir}{Full path to the MultiBUGS install directory}

\item{...}{Passed to \code{\link{bugs_example}}}

\item{opendi}{Full path to the OpenBUGS install directory}
}
\description{
Runs a single Example from the specified MultiBUGS install, with the
specified level of distribution
}
