% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bugs.R
\name{bugs_example_openbugs}
\alias{bugs_example_openbugs}
\title{Run an example in OpenBUGS using R2OpenBUGS}
\usage{
bugs_example_openbugs(model, n.iter = 5000, n.chains = 2,
  n.workers = NULL, dir = "C:/OpenBUGS", examples_dir, working_dir)
}
\arguments{
\item{model}{A character vector (length 1) specifying the model}

\item{n.iter}{The number of iterations. If NULL, the default for each model
is used.}

\item{n.chains}{The number of chains}

\item{n.workers}{The number of cores to use}

\item{dir}{Full path to the MultiBUGS install directory}

\item{examples_dir}{Full path to the directory containing the examples}

\item{working_dir}{Full path to a temp dir where the model should be run}
}
\description{
Run an example in OpenBUGS using R2OpenBUGS
}
