% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{bugs_examples_all}
\alias{bugs_examples_all}
\title{Run all Examples}
\usage{
bugs_examples_all(dir = "C:/MultiBUGS", n.iter = NULL, n.chains = 2,
  n.workers = 2, report = "text", check = "simply_ran",
  exclude = "Sixcomp", include = NULL, save = NULL,
  implementation = "MultiBUGS", examples.dir = file.path(dir,
  "Examples"), ...)
}
\arguments{
\item{dir}{Full path to the MultiBUGS install directory}

\item{n.iter}{The number of iterations. If NULL, the default for each model
is used.}

\item{n.chains}{The number of chains}

\item{n.workers}{The number of cores to use}

\item{report}{Specify how to report results. Either \code{"text"} to use
\code{\link{text_reporter}} or \code{"appveyor"} to use
\code{\link{appveyor_reporter}}}

\item{check}{Specify how to check the results. Either \code{"simply_ran"} to
simply check that the Example ran, or \code{"openbugs"} to check results
against OpenBUGS 3.2.3}

\item{exclude}{A character vector of model names to skip}

\item{include}{A character vector of model names to run. If \code{NULL} then
all models (except those excluded) are run.}

\item{save}{A path where to save results, otherwise \code{NULL}}

\item{implementation}{Either \code{"MultiBUGS"} or \code{"OpenBUGS"}}

\item{examples.dir}{Path to the directory containing the examples}
}
\description{
Runs all Examples in the specified MultiBUGS directory
}
