% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_against_openbugs}
\alias{check_against_openbugs}
\title{Simply check that the model ran}
\usage{
check_against_openbugs(model, output)
}
\arguments{
\item{model}{A character vector (length 1) specifying the model}

\item{output}{The output from running the model. An code{mcmc.list} object
or NULL}
}
\description{
Simply check that the model ran
}
