% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{custom_file}
\alias{custom_file}
\title{Create full paths to arbitrary files in the Examples dir}
\usage{
custom_file(filename, examples_dir)
}
\arguments{
\item{filename}{An arbitrary character string, containing the file name}

\item{examples_dir}{A character vector (length 1), containing the path to
the Examples directory in the MultiBUGS directory}
}
\description{
Create full paths to arbitrary files in the Examples dir
}
