% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arg.R
\name{data_arg}
\alias{data_arg}
\title{Set up data file for running}
\usage{
data_arg(model, examples_dir)
}
\arguments{
\item{model}{A character vector (length 1) specifying the model}

\item{examples_dir}{A character vector (length 1), containing the path to
the Examples directory in the MultiBUGS directory}
}
\value{
The full path to the just-created (as a result of copying) file;
OR an R object with the data
}
\description{
Finds the appropriate data files and either copies or loads the data.
Simple cases will be copied across. More fiddly examples involve loading
the data into R.
}
