% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arg.R
\name{dic_arg}
\alias{dic_arg}
\title{Specify whether to monitor DIC or not}
\usage{
dic_arg(model)
}
\arguments{
\item{model}{A character vector (length 1) specifying the model}
}
\value{
A logical of length 1
}
\description{
Specify whether to monitor DIC or not
}
