% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arg.R
\name{fix_founder_arg}
\alias{fix_founder_arg}
\title{Specify whether founders should be fixed}
\usage{
fix_founder_arg(model)
}
\arguments{
\item{model}{A character vector (length 1) specifying the model}
}
\value{
A logical of length 1
}
\description{
Specify whether founders should be fixed
}
