% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arg.R
\name{inits1_arg}
\alias{inits1_arg}
\title{Set up inits1 file for running}
\usage{
inits1_arg(model, examples_dir)
}
\arguments{
\item{model}{A character vector (length 1) specifying the model}

\item{examples_dir}{A character vector (length 1), containing the path to
the Examples directory in the MultiBUGS directory}
}
\value{
The full path to the just-created (as a result of copying) file
}
\description{
Finds the appropriate inits1 file and either copies it to the current
working directory
}
