% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arg.R
\name{inits_all_arg}
\alias{inits_all_arg}
\title{Set up inits file for running}
\usage{
inits_all_arg(model, examples_dir, n.chains)
}
\arguments{
\item{model}{A character vector (length 1) specifying the model}

\item{examples_dir}{A character vector (length 1), containing the path to
the Examples directory in the MultiBUGS directory}

\item{n.chains}{The number of chains}
}
\value{
The full path to the just-created (as a result of copying) file
}
\description{
Finds the appropriate inits file and either copies it to the current
working directory
}
