% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/junit.R
\name{junit_reporter}
\alias{junit_reporter}
\title{Report result of test for junit}
\usage{
junit_reporter(type, ...)
}
\arguments{
\item{type}{One of \code{"pre"}, \code{"post"} and \code{"wrapup"}
specifying the stage that needs reporting.}
}
\value{
The function \code{\link{junit_reporter_pre}},
\code{\link{junit_reporter_post}} or \code{\link{junit_reporter_wrapup}}
}
\description{
Report result of test for junit
}
