% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/junit.R
\name{junit_reporter_setup}
\alias{junit_reporter_setup}
\title{Report "setup" results to AppVeyor}
\usage{
junit_reporter_setup(dir, n.chains, n.workers, examples_dir)
}
\arguments{
\item{dir}{Full path to the MultiBUGS install directory}

\item{n.chains}{The numebr of chains}

\item{n.workers}{The number of workers used}

\item{examples_dir}{Path to the directory containing the examples}
}
\description{
Report "setup" results to AppVeyor
}
