% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/junit.R
\name{junit_reporter_wrapup}
\alias{junit_reporter_wrapup}
\title{Report "wrapup" results to the console}
\usage{
junit_reporter_wrapup(output_all, passed_all)
}
\arguments{
\item{output_all}{A list of output from all tests}

\item{passed_all}{A logical vector, with TRUE if the test passed}
}
\description{
Report "wrapup" results to the console
}
