% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{load_data_file}
\alias{load_data_file}
\title{Log a simple data file}
\usage{
load_data_file(filename, examples_dir)
}
\arguments{
\item{filename}{A filename}

\item{examples_dir}{A character vector (length 1), containing the path to
the Examples directory in the MultiBUGS directory}
}
\description{
IMPORTANT this does not account for BUGS' row-major ordering!
}
