% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arg.R
\name{model_arg}
\alias{model_arg}
\title{Setup model file for running}
\usage{
model_arg(model, examples_dir)
}
\arguments{
\item{model}{A character vector (length 1) specifying the model}

\item{examples_dir}{A character vector (length 1), containing the path to
the Examples directory in the MultiBUGS directory}
}
\value{
The full path to the just-created (as a result of copying) file
}
\description{
Finds the standard model file for the specified model, and copies it
into the current working directory
}
