% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arg.R
\name{niter_arg}
\alias{niter_arg}
\title{Default n.iter for each model}
\usage{
niter_arg(n.iter, model)
}
\arguments{
\item{n.iter}{Number of iterations, overrides the default}

\item{model}{A character vector (length 1) specifying the model}
}
\description{
Selected to get Gelman-Runbin diagnostic below 1.05
}
