% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arg.R
\name{param_to_save_arg}
\alias{param_to_save_arg}
\title{Specify which params to monitor}
\usage{
param_to_save_arg(model, examples_dir)
}
\arguments{
\item{model}{A character vector (length 1) specifying the model}

\item{examples_dir}{A character vector (length 1), containing the path to
the Examples directory in the MultiBUGS directory}
}
\value{
A character vector of parameter names
}
\description{
By default, any parameter with an init specified is monitored
}
