% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{path_to_filename}
\alias{path_to_filename}
\title{Remove directory separators from file paths}
\usage{
path_to_filename(path)
}
\arguments{
\item{path}{A file path, as a character}
}
\value{
A character string with slashes and colons converted to dashes
}
\description{
Remove directory separators from file paths
}
