% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{standard_copy_and_return}
\alias{standard_copy_and_return}
\title{Copy a standard file to current working directory}
\usage{
standard_copy_and_return(type, model, examples_dir)
}
\arguments{
\item{type}{A character string. One of \code{"model"}, \code{"data"},
\code{"inits"} or \code{"inits1"}}

\item{model}{A character string. The name of one of the Examples, as
returned by \code{\link{all_models_in_dir}} for example.}

\item{examples_dir}{A character vector (length 1), containing the path to
the Examples directory in the MultiBUGS directory}
}
\value{
The path to the just-created file
}
\description{
Copies a standard Examples file to the current working directory and
return the path to the just-created file
}
