% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{standard_file}
\alias{standard_file}
\title{Construct full path to standard Examples files}
\usage{
standard_file(type, model, examples_dir)
}
\arguments{
\item{type}{A character string. One of \code{"model"}, \code{"data"},
\code{"inits"} or \code{"inits1"}}

\item{model}{A character string. The name of one of the Examples, as
returned by \code{\link{all_models_in_dir}} for example.}

\item{examples_dir}{A character vector (length 1), containing the path to
the Examples directory in the MultiBUGS directory}
}
\value{
A character string with the full path to the corresponding file
}
\description{
The Examples in BUGS are named in a standard manner. For a model called
"EXAMPLE", there will be the following files:
}
\details{
1. EXAMPLE.txt - this is the manual page that describes the model, and links
   to other pages
2. EXAMPLEmodel.txt - this just contains the model
3. EXAMPLEdata.txt - this just contains the data
4. EXAMPLEinits.txt - this contains the first set of inits
5. EXAMPLEinits1.txt - this contains a second set of inits

This function will create the full path to each of these from the
components
}
