% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{text_reporter}
\alias{text_reporter}
\title{Report result of test to the console}
\usage{
text_reporter(type, ...)
}
\arguments{
\item{type}{One of \code{"setup"}, \code{"pre"}, \code{"post"} and
\code{"wrapup"} specifying the stage that needs reporting.}
}
\value{
The function \code{\link{text_reporter_setup}},
\code{\link{text_reporter_pre}}, \code{\link{text_reporter_post}} or
\code{\link{text_reporter_wrapup}}
}
\description{
Report result of test to the console
}
