% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{tidy_working_dir}
\alias{tidy_working_dir}
\title{Delete BUGS files from a working_dir}
\usage{
tidy_working_dir(working_dir)
}
\arguments{
\item{working_dir}{A path to the working directory}
}
\description{
Deletes model, inits, inits1 and data from the specified directory
}
